Working test reorganization
