+++ src/container/sequential/io/filemap.rs
@@ -430,3 +430,12 @@ where
         ret
     }
 }
+
+//------------------------------------------------------------------------------
+// Tests
+//------------------------------------------------------------------------------
+
+#[cfg(test)]
+mod tests {
+    use super::FileMap;
+}
