+++ src/container/sequential/mod.rs
@@ -16,6 +16,3 @@ pub use crate::container::sequential::top_k::TopK;
 /// in mediums other than main memory.
 /// ```
 mod io;
-
-#[cfg(test)]
-mod tests;
