+++ src/container/sequential/stack.rs
@@ -242,40 +242,3 @@ where
         self.l1.iter_mut().chain(self.l2.iter_mut())
     }
 }
-
-//------------------------------------------------------------------------------------//
-//                                        Tests                                       //
-//------------------------------------------------------------------------------------//
-
-#[cfg(test)]
-mod tests {
-    use super::Stack;
-    use crate::container::sequential::tests;
-    use crate::container::sequential::Map;
-    use crate::reference::Default;
-
-    #[test]
-    fn test_associative() {
-        tests::test_container(
-            Stack::new(
-                Map::<_, _, Default<_>>::new(10),
-                Map::<_, _, Default<_>>::new(100),
-            ),
-            true,
-        );
-        tests::test_container(
-            Stack::new(
-                Map::<_, _, Default<_>>::new(1),
-                Map::<_, _, Default<_>>::new(10),
-            ),
-            true,
-        );
-        tests::test_container(
-            Stack::new(
-                Map::<_, _, Default<_>>::new(10),
-                Map::<_, _, Default<_>>::new(1),
-            ),
-            true,
-        );
-    }
-}
