+++ src/container/sequential/vector.rs
@@ -214,26 +214,3 @@ where
         self.values.into_iter().map(|(k, r)| (k, r.unwrap()))
     }
 }
-
-//------------------------------------------------------------------------------------//
-//                                        Tests                                       //
-//------------------------------------------------------------------------------------//
-
-#[cfg(test)]
-mod tests {
-    use super::Vector;
-    use crate::container::sequential::tests;
-    use crate::container::tests::test_container;
-
-    #[test]
-    fn container_test() {
-        test_container(Vector::new(10), 100);
-        test_container(Vector::new(100), 10);
-    }
-
-    #[test]
-    fn test_vector() {
-        tests::test_container(Vector::new(10), true);
-        tests::test_container(Vector::new(100), true);
-    }
-}
