+++ src/utils/mod.rs
@@ -1,11 +1,3 @@
-/// Library custom read/write lock.
-pub mod lock;
-
-/// Module for accessing an object concurrently in a mutable state.
-pub mod clone;
-
 pub mod num;
 pub mod ptr;
 pub mod stats;
-
-pub mod rand;
