+++ tests/utils/clone.rs
@@ -1,7 +1,8 @@
-use crate::container::{Concurrent, Container, Iter, IterMut, Sequential};
-use crate::reference::Reference;
-use crate::utils::lock::{RWLock, RWLockGuard};
+use cache::container::{Concurrent, Container, Iter, IterMut, Sequential};
+use cache::lock::{RWLock, RWLockGuard};
+use cache::reference::Reference;
 use std::boxed::Box;
+use std::marker::Sync;
 use std::ops::{Deref, DerefMut, Drop};
 
 //------------------------------------------------------------------------------------//
@@ -27,7 +28,6 @@ impl<V> DerefMut for CloneCell<V> {
 }
 
 impl<V> CloneCell<V> {
-    #[allow(dead_code)]
     pub fn new(value: V) -> Self {
         let rc = RWLock::new();
         // Increment reference count in the lock by one.
@@ -38,14 +38,10 @@ impl<V> CloneCell<V> {
         }
     }
 
-    /// Increment reference count in the lock by one.
     pub fn clone(&self) {
         self.rc.lock();
     }
 
-    /// Function to call when a thread stops using this clone cell.
-    /// The reference count of uses is decreased by one and the
-    /// function returns `true` if the reference count drops to one.
     pub fn drop(&mut self) -> bool {
         self.rc.unlock(); // release (last?) read lock.
         self.rc.try_lock_mut() // Return whether we are the only remaining clone owner.
@@ -87,7 +83,6 @@ impl<V> CloneMut<V> {
     /// CloneMut constructor.
     /// Wraps a value into a `CloneCell`.
     ///
-    #[allow(dead_code)]
     pub fn new(value: V) -> Self {
         CloneMut {
             ptr: Box::into_raw(Box::new(CloneCell::new(value))),
@@ -225,25 +220,5 @@ where
     type Iterator = I;
     fn iter_mut(&'a mut self) -> I {
         unsafe { (*self.ptr).iter_mut() }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::CloneMut;
-    use std::thread;
-
-    #[test]
-    fn test_clone() {
-        let v = CloneMut::new(96u32);
-
-        for i in 0..64 {
-            let mut vc = v.clone();
-            thread::spawn(move || {
-                *vc = i as u32;
-            });
-        }
-
-        assert!(*v < 64u32);
     }
 }
